requirements.txt
Flask==2.0.3
flask-cors
Pillow
rembg